Switch from StableGraph to Graph. Use FxHashMap instead of Hashmap.
Reason for the switch is because StableGraph doesn't actually clear indices for nodes and edges. This means there's a cap on the total number of nodes/edges in the program's lifetime, which isn't what we want with a long running client. To alleviate this, Graph is used and aml_graph keeps track of indices to ensure the index maps don't break for nodes/edges.